% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_measure.R
\name{comparison_measure}
\alias{comparison_measure}
\alias{RMSE}
\alias{MAE}
\alias{MSE}
\title{Comparative Measures for Regression Models}
\usage{
comparison_measure(model, type = c("auto", "liner", "power"))

RMSE(model, type = c("auto", "liner", "power"))

MAE(model, type = c("auto", "liner", "power"))

MSE(model, type = c("auto", "liner", "power"))
}
\arguments{
\item{model}{A linear regression model or power regression model of the lm class created by the \code{lm()} function.}

\item{type}{Selects whether the regression model is a linear regression model or a power regression model.
When set to auto, it determines whether the dependent variable contains a log transformation and selects accordingly.
See note.}
}
\description{
A function to calculate comparison metrics for regression models, such as RMSE, MAE, and MSE.
\code{comparison_measure} displays all three metrics together.
Each value can be calculated using its respective function.
Each definition is based on Tarald O. Kvalseth's (1985) work.
Particular attention should be paid to MSE. See details.
}
\details{
The root mean squared residual or error (RMSE), mean absolute residual (MAE), and the mean squared residual (MSE) defined as based on Kvålseth (1985).
\deqn{\text{RMSE} = [{\frac{\sum (y - \hat{y})^2}{n}}]^{1/2}}
\deqn{\text{MAE} = \frac{\sum |y - \hat{y}|}{n}}
\deqn{\text{MSE} = \frac{\sum (y - \hat{y})^2}{n - p}}
where \eqn{p} denotes the number of model parameters.
Note that the MSE generally used today is not adjusted for the number of parameters, whereas Kvalseth's (1985) MSE is adjusted for the number of parameters.
}
\note{
The power regression model must be based on a logarithmic transformation.

The auto-selection between linear regression and power regression models is determined by whether the dependent variable's name contains “log”.
If the name “log” is intentionally used for a linear regression model, the selection cannot be made correctly.
}
\examples{
# example data set 1. Kvålseth (1985).
df1 <- data.frame(x = c(1:6),
                 y = c(15,37,52,59,83,92))
model_intercept <- lm(y ~ x, df1)
model_without <- lm(y ~ x - 1, df1)
model_power <- lm(log(y) ~ log(x), df1)
comparison_measure(model_intercept)
comparison_measure(model_without)
comparison_measure(model_power)
}
\references{
Tarald O. Kvalseth (1985) Cautionary Note about R 2 , The American Statistician, 39:4, 279-285, \doi{DOI: 10.1080/00031305.1985.10479448}
}
