% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kvr2.R
\name{plot.r2_kvr2}
\alias{plot.r2_kvr2}
\title{Plot Method for r2_kvr2 Objects}
\usage{
\method{plot}{r2_kvr2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{r2_kvr2}.}

\item{...}{Currently ignored.}
}
\value{
A \code{ggplot} object representing the visual analysis.
}
\description{
Visualizes the nine definitions of R-squared to compare their values
and identify potential issues (e.g., values exceeding 1 or falling below 0).
}
\examples{
df1 <- data.frame(x = 1:6, y = c(15, 37, 52, 59, 83, 92))
model <- lm(y ~ x - 1, data = df1) # No-intercept model
r2(model)

}
