% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kvr2.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plot Observed vs Predicted Values}
\usage{
plot_diagnostic(x, ...)
}
\arguments{
\item{x}{A fitted \code{lm} object.}

\item{...}{Currently ignored.}
}
\value{
A \code{ggplot} object representing the visual analysis.
}
\description{
A diagnostic plot to visualize why R-squared might be low or negative.
It compares the model predictions (identity line) against the mean (horizontal line).
}
\examples{
df1 <- data.frame(x = 1:6, y = c(15, 37, 52, 59, 83, 92))
model <- lm(y ~ x - 1, data = df1) # No-intercept model
plot_diagnostic(model)

}
