% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_r2.R
\name{print.r2}
\alias{print.r2}
\alias{print.comparison_measure}
\title{Print Methods for r2 and comparison_measure calculation Objects}
\usage{
\method{print}{r2}(x, ..., digits = 4)

\method{print}{comparison_measure}(x, ..., digits = 4)
}
\arguments{
\item{x}{description object of class "\code{r2}" or "\code{comparison_measure}".}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{number of significant digits to be used.}
}
\description{
Printing objects of class "\code{r2}" or "\code{comparison_measure}", respectively, by simple print methods.
}
\seealso{
\link{r2} \link{comparison_measure} \link{r2_adjusted}
}
